# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "A-Share Stock Backtesting & Quant Platform API",
    "version": "1.0.0",
    "description": "A-share 量化平台后端 REST API。\n\n主要业务领域：\n- **认证 & 用户**：登录/注册/JWT 刷新\n- **回测 & 策略**：单股回测、组合级策略、参数优化\n- **因子**：因子库管理、IC 报告、相关性矩阵\n- **模拟交易**：持仓、订单、风控守卫\n- **量化引擎**：StrategyEngine / SignalEngine / BacktestEngine 等 5 大 facade\n- **数据 & 行情**：股票元信息、K 线、实时行情、龙虎榜、北向资金\n- **AI**：TradingAgents 深度分析、KOL 观点、公告 NLP、市场情绪\n- **推送 & 告警**：飞书 / 邮件 / 微信 / 短信 4 channel\n- **运维 & 监控**：env 校验、CI 状态、健康度检查\n\n所有 endpoint 默认返回 `{ success: boolean, data?: any, message?: string }` envelope。\n需要鉴权的 endpoint 需在 Header 携带 `Authorization: Bearer <JWT>`。",
    "contact": {
      "name": "QuantX A-Share Alpha Refactor"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development (本地)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在 Authorization Header 携带 `Bearer <JWT>`。Token 通过 `/api/auth/login` 获取。"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "业务数据（结构因 endpoint 而异）"
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/paper-trading": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取当前用户的模拟盘数据及持仓明细",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "组合与持仓数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/trade": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "在模拟盘中进行买入或卖出交易",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "symbol": {
                    "type": "string"
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "buy",
                      "sell"
                    ]
                  },
                  "quantity": {
                    "type": "number"
                  },
                  "price": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成交结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/auto-from-signals": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "从已归档 AI/量化推荐信号自动生成模拟盘交易",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "自动建仓结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/auto-sync-recommendations": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "刷新候选推荐、归档为投研信号并自动进入模拟盘",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "同步结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/refresh-snapshot": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "刷新模拟盘最新价格与资金快照",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "最新快照"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/risk-check": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "按止损/止盈/卖出信号/最长持有期检查并自动退出",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "风控检查结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/autonomous-dashboard": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取自主荐股模拟盘收益驾驶舱",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "驾驶舱数据"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/autonomous-auto-sync": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "自主荐股闭环 - 推荐、归档信号并进入 20W 自主模拟盘",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "闭环结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/autonomous-risk-check": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "自主荐股闭环 - 按卖出信号/止损/止盈/持有期结算 20W 自主模拟盘",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "结算结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/recommendation-tracking": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取每日推荐股票追踪与模拟收益",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "推荐追踪数据"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/autonomous-optimization": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取自主荐股闭环优化台（策略晋级/收益路径/降权放大片段）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "优化台数据"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/attribution": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取模拟盘信号收益归因与策略反哺",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "归因数据"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/risk-profile": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取模拟盘组合风险画像（现金水位/总仓位/回撤/集中度/相关性/VaR）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "风险画像数据"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/order-intents": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取模拟交易订单意图与拒单归因",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "订单意图列表"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/order-intents/family-hindsight": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取全部策略账户拒单后验汇总（错杀/有效拦截/账户级规则建议）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "后验汇总"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/order-intents/{id}/trace": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取单条订单意图链路（信号/拒单原因/后验收益/规则建议/参数影响）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "链路详情"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "未找到"
          }
        }
      }
    },
    "/api/paper-trading/order-intents/hindsight/refresh": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "主动刷新订单意图后验快照",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "刷新结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/recommendation-outcomes": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取推荐信号到模拟交易收益的闭环看板",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "闭环看板"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/recommendation-outcomes/{id}/trace": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取单笔推荐链路详情（信号/复核/风控/模拟交易/收益）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "链路详情"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "未找到"
          }
        }
      }
    },
    "/api/paper-trading/recommendation-outcomes/refresh": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "刷新推荐信号到模拟交易收益的闭环结果",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "刷新结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/recommendation-outcomes/report": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "将推荐交易收益闭环报告写入飞书多维表格",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "写入结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/attribution/report": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "生成模拟盘收益归因并写入飞书多维表格",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "写入结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/plan": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "生成模拟盘盘前/盘后交易计划",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "交易计划"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/plan/report": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "生成模拟盘交易计划并写入飞书多维表格",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "写入结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/order-intent-tuning/apply": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "预览或手动应用订单意图稳定窗口给出的调参建议",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "应用结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/order-intent-tuning/candidates": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取订单意图调参只读候选（稳定窗口 + 多账户拒单后验）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "候选列表"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/order-intent-tuning/canary/snapshots": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取订单意图 Canary 评审快照时间线",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "快照时间线"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/order-intent-tuning/canary": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取订单意图 Canary 小流量调参观察状态",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "观察状态"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/order-intent-tuning/canary/rollback": {
      "post": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "预览或强确认回滚订单意图 Canary 小流量调参",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "回滚结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/history": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取模拟盘的交易流水",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "交易流水列表"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/snapshots": {
      "get": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "获取模拟盘的资金曲线快照",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "资金曲线快照"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
    "/api/paper-trading/positions/{id}/stop-loss": {
      "put": {
        "tags": [
          "模拟交易 PaperTrading"
        ],
        "summary": "设置/清除指定持仓的硬止损价 (US-017)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "stop_loss_price": {
                    "type": "number",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "设置结果"
          },
          "400": {
            "description": "参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "未找到持仓"
          }
        }
      }
    },
...
  }
}